# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JobTrackr API",
        "description": "API documentation for JobTrackr - A job application tracking platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get all user applications",
                "operationId": "dd67feb5de774e8f7106ff3ea81f575c",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Applied",
                                "Interview",
                                "Technical Test",
                                "Offer",
                                "Rejected"
                            ]
                        }
                    },
                    {
                        "name": "company",
                        "in": "query",
                        "description": "Search by company name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by field (applied_date, company, position)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of applications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Application"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Create a new application",
                "operationId": "19f9c4e2b9bb3e982db6569cf02a59d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company",
                                    "position",
                                    "applied_date"
                                ],
                                "properties": {
                                    "company": {
                                        "type": "string",
                                        "example": "Google"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Senior Developer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Applied",
                                            "Interview",
                                            "Technical Test",
                                            "Offer",
                                            "Rejected"
                                        ]
                                    },
                                    "job_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "job_description": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "applied_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-14"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications/{id}": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get a specific application",
                "operationId": "1068983793babb1ce7b105cee2aeade4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Application not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Applications"
                ],
                "summary": "Update an application",
                "operationId": "30371963d7c7e9a044faab88583b7809",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "company": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Applied",
                                            "Interview",
                                            "Technical Test",
                                            "Offer",
                                            "Rejected"
                                        ]
                                    },
                                    "job_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "job_description": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "applied_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Application updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Applications"
                ],
                "summary": "Delete an application",
                "operationId": "45bc43746912d036435136a505aa1ffc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Application": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "company": {
                        "type": "string",
                        "example": "Google"
                    },
                    "position": {
                        "type": "string",
                        "example": "Senior Developer"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Applied",
                            "Interview",
                            "Technical Test",
                            "Offer",
                            "Rejected"
                        ]
                    },
                    "job_url": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    },
                    "job_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "applied_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "documents_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}